dialog fixes
changed most, if not all dialog to first person past tense. removed the time-based occurences, such as the fireplace prints, and replaced them with static pathways or a simple flag raised by a different action to give the illusion of time passing.
Closer examination is needed to ensure that everything works and to make sure no pathways have present tense or second-person narration. For this purpose, I have also updated the Proofing Copy.